# docs/api-spec/openapi.yaml
openapi: 3.0.3
info:
  title: NewArch API
  description: NewArch 微服务平台 API 文档
  version: 1.0.0
  contact:
    name: NewArch Team
servers:
  - url: http://localhost:8080/api/v1
    description: 本地开发环境
...
paths:
...
  /tasks:
    get:
      tags: [Task]
      summary: 列出任务
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, running, completed, failed, cancelled]
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  tasks:
                    type: array
                    items:
                      $ref: '#/components/schemas/Task'
    post:
      tags: [Task]
      summary: 创建任务
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreate'
      responses:
        '201':
          description: 创建成功
  /tasks/{task_id}:
    get:
      tags: [Task]
      summary: 获取任务详情
      security:
        - bearerAuth: []
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
    delete:
      tags: [Task]
      summary: 取消任务
      security:
        - bearerAuth: []
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 取消成功

  # ==================== AI ====================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Task:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        type:
          type: string
        status:
          type: string
        title:
          type: string
        progress:
          type: integer
        created_at:
          type: string
          format: date-time
    TaskCreate:
      type: object
      required: [type, title]
      properties:
        type:
          type: string
          enum: [browse, search, extract, analyze, custom]
        title:
          type: string
        description:
          type: string
        input:
          type: object
...